Fix Special Field documentation
The example in the documentation is currently not working and using deprecated syntax according to Vue's documentation (https://vuejs.org/v2/guide/components-slots.html#Deprecated-Syntax).

I got it to work by adding a `__slot:` prefix in the field's name. Additionally, the non-deprecated slot syntax is also a lot more concise and supports destructuring as well. I incorporated that as well. I was not sure whether dashes are allowed in slot names, so I just changed the name to `genderSlot` just to be sure.